Save object and findPtr
findPtr works - as far as I can see

save object needs some rethinking. It needs to make stable ptr to the
object as otherwise it just gets moved around by the gc and the address
is garbage.

Another option is to allow the client to transmit an address when they
pause.