Add support for forking debuggee process
This allows the main process to keep running as normal whilst the child
process is paused to service the debugger. When the child process
attempts to resume then if it has been forked then it exits and passed
control back to the parent, which waits until the child has finished
before proceding.