Fix parralel tracing function
The previous implementation relied on BlockedIndefinitelyOnMVar
exception, which is not suitable for the control flow of the program as
it's unpredicatable when the RTS will throw it.

Therefore we need to explicitly look when each chan is empty, and there
is no more work for when to finish.

In order to do this I changed the implementatin to use a TChan and then
there's a blocking atomic block which queries each TChan in order to see
if it's finished or not.

Core utilisation doesn't look too great. It seemed to only use about 50%
of each capability when looking in htop.