Add support for SRTs
SRTs describe how thunks, functions and stack frames retain static
objects. Previously we didn't trace these so any leaks due to retaining
static objects would be very hard to track down. Now those pointers are
added into the graph and a new request type (RequestSRT) is added for
decoding the SRT for an info table.

A consequence is that there's a new type parameter for the DebugClosure
type which describes whether we have decoded the SRT field yet for a
closure.