Add support for decoding profiling headers
Query debugee for profiling mode and TNTC

Deserialize closures instead of relying on ghc-heap
The above is necessary to decode profiling headers and ghc-heap
does not

Additionally, deserialize info tables in haskell code instead of
relying on ghc-heap. This is also necessary to decode profiling
info tables properly.

Perhaps we should introduce a type for profiling info tables instead of
using StgInfoTable